requirements.txt
pandas==2.2.2
numpy==1.26.4
pyyaml==6.0.1
matplotlib==3.8.2
seaborn==0.12.2
